# api/v1/openapi.yaml
openapi: 3.0.0
info:
  title: Fungicibus Order Service API
  description: API for managing orders
  version: 1.0.0

servers:
  - url: /api/v1

components:
  schemas:
    # Error
    ErrorResponse:
      type: object
      required:
        - errors
      properties:
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ErrorItem"
    ErrorItem:
      type: object
      required:
        - id
        - status
        - detail
        - meta
      properties:
        id:
          type: string
          description: Identificator of a request
          example: bd124f62-27ff-4ed7-ad8b-5df0fa73e2c1
        status:
          type: string
          description: The HTTP status code applicable to this problem, expressed as a string value
          example: "400"
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence of the problem
          example: Bad commodity name
        source:
          $ref: "#/components/schemas/ErrorSource"
        meta:
          $ref: "#/components/schemas/ErrorMeta"
    ErrorSource:
      type: object
      required:
        - pointer
      properties:
        pointer:
          type: string
          description: a JSON Pointer [RFC6901] to the value in the request document that caused the error
          example: /data/quantity
    ErrorMeta:
      type: object
      properties:
        timestamp:
          type: string
          description: Timestamp of the occurrence of the problem in RFC3339 format
          example: 2025-01-02T15:04:05Z07:00

    CreateOrderRequest:
      type: object
      required:
        - data
      properties:
        data:
          $ref: "#/components/schemas/CreateOrderItem"

    CreateOrderItem:
      type: object
      required:
        - productId
        - quantity
        - timestamp
      properties:
        productId: 
          type: string
          description: Id of a product
          example: "3a550dd9-c7d4-4c77-a73a-fcd13ce76712"
        quantity: 
          type: integer
          description: Quantity of a product
          example: 1
        timestamp: 
          type: string
          description: Timestamp of an order in RFC3339 format
          example: "2006-01-02T15:04:05"
        comment: 
          type: string
          description: Comment to an order
          example: "Some comment"

    CreateOrderResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: "#/components/schemas/CreatedOrder"

    CreatedOrder:
      type: object
      allOf:
        - $ref: '#/components/schemas/CreateOrderItem' 
        - type: object
          properties:
            id:
              type: string
              description: Id of an order
              example: "3a550dd9-c7d4-4c77-a73a-fcd13ce76712"
          required:
            - id
paths:
  /orders:
    post:
      summary: Create order
      operationId: CreateOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateOrderResponse"
        "400":
          description: Request error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"